backend/requirements.txt
python-multipart==0.0.12
Hypercorn[uvloop]==0.17.3
fastapi==0.115.0
google-api-python-client==2.147.0
google-cloud-logging==3.11.2
google-cloud-secret-manager==2.20.2
google-cloud-speech==2.27.0
google-cloud-texttospeech==2.17.2
google-generativeai==0.8.2
firebase-admin==6.5.0
python-dotenv==1.0.1
redis==5.0.8